test: move export test to a better location
