+++ util/test/exports.js
@@ -1,0 +1,44 @@
+import test from 'tape-catch';
+
+import candela from '../../candela';
+
+function structureTests (t, cd) {
+  t.ok(cd, 'candela exists');
+
+  t.ok(cd.components, 'candela.components exists');
+  t.equal(typeof cd.components, 'object', 'candela.components is an object');
+
+  t.ok(cd.register, 'candela.register exists');
+  t.equal(typeof cd.register, 'function', 'candela.register is a function');
+
+  t.ok(cd.unregister, 'candela.unregister exists');
+  t.equal(typeof cd.unregister, 'function', 'candela.unregister is a function');
+
+  t.ok(cd.unregisterAll, 'candela.unregisterAll exists');
+  t.equal(typeof cd.unregisterAll, 'function', 'candela.unregisterAll is a function');
+
+  t.ok(cd.VisComponent, 'candela.VisComponent exists');
+  t.equal(typeof cd.VisComponent, 'function', 'candela.VisComponent is a function');
+
+  t.equal(Object.keys(cd).length, 5, 'candela contains no other members');
+}
+
+test('Structure and content of exported Candela library object', t => {
+  structureTests(t, candela);
+  t.end();
+});
+
+test('Structure and content of unminified Candela library file', t => {
+  try {
+    let candelaBuilt = require('../../dist/candela.js');
+    t.ok(candelaBuilt, 'The built candela.js file exists');
+
+    if (candelaBuilt) {
+      structureTests(t, candelaBuilt);
+    }
+  } catch (e) {
+    t.skip('Built candela.js file does not exist');
+  }
+
+  t.end();
+});
